Moved do_shell(), do_edit(), and do_loadsql() into do_func class
